Fix CI: move marked-gfm-heading-id types to standalone ambient file
The declare module in local-types/index.d.ts was treated as a module
augmentation (not an ambient declaration) because index.d.ts has
top-level imports making it a module. Ambient declarations must live
in script files with no top-level import/export.

Created a separate marked-gfm-heading-id.d.ts with only the declare
module block (using import() type expressions to avoid top-level
imports), following the same pattern as matrix-js-sdk.d.ts.
