+++ packages/local-types/index.d.ts
@@ -119,16 +119,6 @@ declare module '@ember/routing' {
 // incorporate
 import '../../runtime-common/global';
 
-declare module 'marked-gfm-heading-id' {
-  import type { MarkedExtension } from 'marked';
-  export function gfmHeadingId(options?: { prefix?: string }): MarkedExtension;
-  export interface HeadingData {
-    level: number;
-    text: string;
-    id: string;
-  }
-  export function getHeadingList(): HeadingData[];
-  export function resetHeadings(): void;
-}
+import './marked-gfm-heading-id';
 
 import './matrix-js-sdk';
