+++ packages/local-types/marked-gfm-heading-id.d.ts
@@ -1,0 +1,12 @@
+declare module 'marked-gfm-heading-id' {
+  export function gfmHeadingId(
+    options?: { prefix?: string },
+  ): import('marked').MarkedExtension;
+  export interface HeadingData {
+    level: number;
+    text: string;
+    id: string;
+  }
+  export function getHeadingList(): HeadingData[];
+  export function resetHeadings(): void;
+}
