+++ packages/runtime-common/global.d.ts
@@ -12,15 +12,3 @@ declare module '@babel/plugin-transform-typescript' {
   import type * as Babel from '@babel/core';
   export default function makePlugin(babel: typeof Babel): Babel.PluginObj;
 }
-
-declare module 'marked-gfm-heading-id' {
-  import type { MarkedExtension } from 'marked';
-  export function gfmHeadingId(options?: { prefix?: string }): MarkedExtension;
-  export interface HeadingData {
-    level: number;
-    text: string;
-    id: string;
-  }
-  export function getHeadingList(): HeadingData[];
-  export function resetHeadings(): void;
-}
